+++ lib/active_model/serializer/collection_reflection.rb
@@ -1,7 +1,0 @@
-module ActiveModel
-  class Serializer
-    # @api private
-    class CollectionReflection < Reflection
-    end
-  end
-end