+++ lib/active_model/serializer/concerns/caching.rb
@@ -40,9 +40,9 @@ module ActiveModel
 
       module ClassMethods
         def inherited(base)
-          super
           caller_line = caller[1]
           base._cache_digest_file_path = caller_line
+          super
         end
 
         def _cache_digest
@@ -66,6 +66,18 @@ module ActiveModel
 
         def _skip_digest?
           _cache_options && _cache_options[:skip_digest]
+        end
+
+        # @api private
+        # maps attribute value to explicit key name
+        # @see Serializer::attribute
+        # @see Serializer::fragmented_attributes
+        def _attributes_keys
+          _attributes_data
+            .each_with_object({}) do |(key, attr), hash|
+            next if key == attr.name
+            hash[attr.name] = { key: key }
+          end
         end
 
         def fragmented_attributes
@@ -158,6 +170,7 @@ module ActiveModel
 
         # Read cache from cache_store
         # @return [Hash]
+        # Used in CollectionSerializer to set :cached_attributes
         def cache_read_multi(collection_serializer, adapter_instance, include_directive)
           return {} if ActiveModelSerializers.config.cache_store.blank?
 
@@ -180,12 +193,14 @@ module ActiveModel
             cache_keys << object_cache_key(serializer, adapter_instance)
 
             serializer.associations(include_directive).each do |association|
-              if association.serializer.respond_to?(:each)
-                association.serializer.each do |sub_serializer|
+              # TODO(BF): Process relationship without evaluating lazy_association
+              association_serializer = association.lazy_association.serializer
+              if association_serializer.respond_to?(:each)
+                association_serializer.each do |sub_serializer|
                   cache_keys << object_cache_key(sub_serializer, adapter_instance)
                 end
               else
-                cache_keys << object_cache_key(association.serializer, adapter_instance)
+                cache_keys << object_cache_key(association_serializer, adapter_instance)
               end
             end
           end
@@ -203,23 +218,18 @@ module ActiveModel
 
       ### INSTANCE METHODS
       def fetch_attributes(fields, cached_attributes, adapter_instance)
-        if serializer_class.cache_enabled?
-          key = cache_key(adapter_instance)
-          cached_attributes.fetch(key) do
-            serializer_class.cache_store.fetch(key, serializer_class._cache_options) do
-              attributes(fields, true)
-            end
+        key = cache_key(adapter_instance)
+        cached_attributes.fetch(key) do
+          fetch(adapter_instance, serializer_class._cache_options, key) do
+            attributes(fields, true)
           end
-        elsif serializer_class.fragment_cache_enabled?
-          fetch_attributes_fragment(adapter_instance, cached_attributes)
-        else
-          attributes(fields, true)
         end
       end
 
-      def fetch(adapter_instance, cache_options = serializer_class._cache_options)
+      def fetch(adapter_instance, cache_options = serializer_class._cache_options, key = nil)
         if serializer_class.cache_store
-          serializer_class.cache_store.fetch(cache_key(adapter_instance), cache_options) do
+          key ||= cache_key(adapter_instance)
+          serializer_class.cache_store.fetch(key, cache_options) do
             yield
           end
         else
@@ -230,7 +240,6 @@ module ActiveModel
       # 1. Determine cached fields from serializer class options
       # 2. Get non_cached_fields and fetch cache_fields
       # 3. Merge the two hashes using adapter_instance#fragment_cache
-      # rubocop:disable Metrics/AbcSize
       def fetch_attributes_fragment(adapter_instance, cached_attributes = {})
         serializer_class._cache_options ||= {}
         serializer_class._cache_options[:key] = serializer_class._cache_key if serializer_class._cache_key
@@ -239,22 +248,21 @@ module ActiveModel
         non_cached_fields = fields[:non_cached].dup
         non_cached_hash = attributes(non_cached_fields, true)
         include_directive = JSONAPI::IncludeDirective.new(non_cached_fields - non_cached_hash.keys)
-        non_cached_hash.merge! resource_relationships({}, { include_directive: include_directive }, adapter_instance)
+        non_cached_hash.merge! associations_hash({}, { include_directive: include_directive }, adapter_instance)
 
         cached_fields = fields[:cached].dup
         key = cache_key(adapter_instance)
         cached_hash =
           cached_attributes.fetch(key) do
-            serializer_class.cache_store.fetch(key, serializer_class._cache_options) do
+            fetch(adapter_instance, serializer_class._cache_options, key) do
               hash = attributes(cached_fields, true)
               include_directive = JSONAPI::IncludeDirective.new(cached_fields - hash.keys)
-              hash.merge! resource_relationships({}, { include_directive: include_directive }, adapter_instance)
+              hash.merge! associations_hash({}, { include_directive: include_directive }, adapter_instance)
             end
           end
         # Merge both results
         adapter_instance.fragment_cache(cached_hash, non_cached_hash)
       end
-      # rubocop:enable Metrics/AbcSize
 
       def cache_key(adapter_instance)
         return @cache_key if defined?(@cache_key)