+++ lib/active_model/serializer/has_many_reflection.rb
@@ -1,7 +1,10 @@
 module ActiveModel
   class Serializer
     # @api private
-    class HasManyReflection < CollectionReflection
+    class HasManyReflection < Reflection
+      def collection?
+        true
+      end
     end
   end
 end